Trekk bruk av registerAktivitet ut av context
