+++ src/frontend/context/InngangsvilkårContext.ts
@@ -1,4 +1,4 @@
-import React, { SetStateAction, useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 
 import constate from 'constate';
 
@@ -6,7 +6,6 @@ import { Aktivitet } from '../Sider/Behandling/Inngangsvilkår/typer/aktivitet';
 import { Målgruppe } from '../Sider/Behandling/Inngangsvilkår/typer/målgruppe';
 import { Stønadsperiode } from '../Sider/Behandling/Inngangsvilkår/typer/stønadsperiode';
 import { Vilkårperioder } from '../Sider/Behandling/Inngangsvilkår/typer/vilkårperiode';
-import { Registeraktivitet } from '../typer/registeraktivitet';
 
 interface UseInngangsvilkår {
     målgrupper: Målgruppe[];
@@ -19,11 +18,6 @@ interface UseInngangsvilkår {
     stønadsperiodeFeil: string | undefined;
     settStønadsperiodeFeil: (feilmelding: string | undefined) => void;
     oppdaterStønadsperioder: (oppdaterteStønadsperioder: Stønadsperiode[]) => void;
-    leggTilAktivitetFraRegister: (aktivitet: Registeraktivitet) => void;
-    leggerTilNyAktivitet: boolean;
-    settLeggerTilNyAktivitet: React.Dispatch<SetStateAction<boolean>>;
-    aktivitetFraRegister?: Registeraktivitet;
-    settAktivitetFraRegister: React.Dispatch<SetStateAction<Registeraktivitet | undefined>>;
 }
 
 interface Props {
@@ -35,7 +29,6 @@ export const [InngangsvilkårProvider, useInngangsvilkår] = constate(
     ({ vilkårperioder, hentedeStønadsperioder }: Props): UseInngangsvilkår => {
         const [målgrupper, settMålgrupper] = useState<Målgruppe[]>(vilkårperioder.målgrupper);
         const [aktiviteter, settAktiviteter] = useState<Aktivitet[]>(vilkårperioder.aktiviteter);
-        const [leggerTilNyAktivitet, settLeggerTilNyAktivitet] = useState<boolean>(false);
         const [stønadsperioder, settStønadsperioder] =
             useState<Stønadsperiode[]>(hentedeStønadsperioder);
         const [stønadsperiodeFeil, settStønadsperiodeFeil] = useState<string>();
@@ -60,15 +53,6 @@ export const [InngangsvilkårProvider, useInngangsvilkår] = constate(
             settAktiviteter((prevState) => [...prevState, nyPeriode]);
         };
 
-        const leggTilAktivitetFraRegister = (aktivitet: Registeraktivitet) => {
-            if (leggerTilNyAktivitet) return;
-
-            settAktivitetFraRegister(aktivitet);
-            settLeggerTilNyAktivitet(true);
-        };
-
-        const [aktivitetFraRegister, settAktivitetFraRegister] = useState<Registeraktivitet>();
-
         const oppdaterAktivitet = (oppdatertPeriode: Aktivitet) => {
             settAktiviteter((prevState) =>
                 prevState.map((aktivitet) =>
@@ -87,13 +71,8 @@ export const [InngangsvilkårProvider, useInngangsvilkår] = constate(
             stønadsperioder,
             stønadsperiodeFeil,
             settStønadsperiodeFeil,
-            leggTilAktivitetFraRegister,
             oppdaterStønadsperioder: (oppdaterteStønadsperioder: Stønadsperiode[]) =>
                 settStønadsperioder(oppdaterteStønadsperioder),
-            leggerTilNyAktivitet,
-            settLeggerTilNyAktivitet,
-            aktivitetFraRegister,
-            settAktivitetFraRegister,
         };
     }
 );
